tests

+++ test/MasserOne.guard.t.sol
@@ -16,7 +16,7 @@ contract MaseerOneGuardTest is MaseerTestBase {
     function testBadActors() public view {
         // Run through the list of bad actors and ensure they are blocked
         uint256 _len = ofac.length();
-        for (uint i = 0; i < _len; i++) {
+        for (uint256 i = 0; i < _len; i++) {
             assertEq(maseerOne.canPass(ofac.ofac(i)), false);
         }
     }
